Rename web/realtime package to web/sockets
