+++ cmd/mailroom/main.go
@@ -23,8 +23,8 @@ import (
 	_ "github.com/nyaruka/mailroom/v26/web/org"
 	_ "github.com/nyaruka/mailroom/v26/web/po"
 	_ "github.com/nyaruka/mailroom/v26/web/public"
-	_ "github.com/nyaruka/mailroom/v26/web/realtime"
 	_ "github.com/nyaruka/mailroom/v26/web/simulation"
+	_ "github.com/nyaruka/mailroom/v26/web/sockets"
 	_ "github.com/nyaruka/mailroom/v26/web/system"
 	_ "github.com/nyaruka/mailroom/v26/web/ticket"
 )
