+++ web/sockets/index_test.go
@@ -1,4 +1,4 @@
-package realtime
+package sockets
 
 import (
 	"testing"
